+++ src/platforms/arm/stm32/pins/boards/f1/generic_f103c8.h
@@ -7,6 +7,13 @@
 //
 // Arduino-style pin numbering for STM32duino
 
+// IWYU pragma: private, include "platforms/arm/stm32/pins/families/stm32f1.h"
+
+#include "platforms/arm/stm32/pins/core/armpin_template.h"
+#include "platforms/arm/stm32/pins/core/gpio_port_init.h"
+#include "platforms/arm/stm32/pins/core/pin_macros.h"
+#include "platforms/arm/stm32/pins/families/stm32f1.h"
+
 #define MAX_PIN 36
 
 // PA0-PA15