+++ src/platforms/arm/stm32/pins/boards/f4/f411re_nucleo.h
@@ -6,6 +6,13 @@
 // Defines pin mappings inside fl:: namespace
 // STM32F411RET6 - 51 pins
 
+// IWYU pragma: private, include "platforms/arm/stm32/pins/families/stm32f4.h"
+
+#include "platforms/arm/stm32/pins/core/armpin_template.h"
+#include "platforms/arm/stm32/pins/core/gpio_port_init.h"
+#include "platforms/arm/stm32/pins/core/pin_macros.h"
+#include "platforms/arm/stm32/pins/families/stm32f4.h"
+
 // Port A
 _DEFPIN_ARM_F4(0, 3, A);   // PA_3
 _DEFPIN_ARM_F4(1, 2, A);   // PA_2